Fix perfomance issue with lots of appointments
