+++ testing/tests/DevExpress.ui.widgets.scheduler/layoutManager.tests.js
@@ -6,8 +6,6 @@ import $ from "jquery";
 import keyboardMock from "../../helpers/keyboardMock.js";
 import pointerMock from "../../helpers/pointerMock.js";
 import translator from "animation/translator";
-import fx from "animation/fx";
-import { DataSource } from "data/data_source/data_source";
 
 import SchedulerLayoutManager from "ui/scheduler/ui.scheduler.appointments.layout_manager";
 import BaseAppointmentsStrategy from "ui/scheduler/rendering_strategies/ui.scheduler.appointments.strategy.base";
@@ -51,322 +49,6 @@ const moduleOptions = {
     afterEach: function() {
     }
 };
-
-const renderLayoutModuleOptions = {
-    beforeEach: function() {
-        this.clock = sinon.useFakeTimers();
-        fx.off = true;
-    },
-    afterEach: function() {
-        this.clock.restore();
-        fx.off = false;
-    }
-};
-
-
-QUnit.module("Render layout", renderLayoutModuleOptions, function() {
-    const createScheduler = (view, dataSource, options) => {
-        const instance = $("#scheduler").dxScheduler($.extend(options, {
-            views: ["week", "month", "agenda"],
-            currentView: view,
-            dataSource: dataSource,
-            currentDate: new Date(2017, 4, 25),
-            startDayHour: 9,
-            height: 600,
-            width: 1300,
-            editing: true,
-        })).dxScheduler("instance");
-
-        return new SchedulerTestWrapper(instance);
-    };
-
-    const markAppointments = (scheduler) => scheduler.appointments.getAppointments().data("mark", true);
-
-    const getUnmarkedAppointments = (scheduler) => {
-        return scheduler.appointments.getAppointments().filter(function() {
-            return !!$(this).data("mark") === false;
-        });
-    };
-
-    const defaultData = [
-        {
-            id: 0,
-            text: "Website Re-Design Plan",
-            startDate: new Date(2017, 4, 21, 9, 30),
-            endDate: new Date(2017, 4, 21, 11, 30)
-        }, {
-            id: 1,
-            text: "Install New Database",
-            startDate: new Date(2017, 4, 21, 12, 45),
-            endDate: new Date(2017, 4, 21, 13, 15)
-        }, {
-            id: 2,
-            text: "Book Flights to San Fran for Sales Trip",
-            startDate: new Date(2017, 4, 22, 12, 0),
-            endDate: new Date(2017, 4, 22, 13, 0),
-        }, {
-            id: 3,
-            text: "Install New Router in Dev Room",
-            startDate: new Date(2017, 4, 22, 14, 30),
-            endDate: new Date(2017, 4, 22, 15, 30)
-        }, {
-            id: 4,
-            text: "Approve Personal Computer Upgrade Plan",
-            startDate: new Date(2017, 4, 23, 10, 0),
-            endDate: new Date(2017, 4, 23, 11, 0)
-        }, {
-            id: 5,
-            text: "Final Budget Review",
-            startDate: new Date(2017, 4, 23, 12, 0),
-            endDate: new Date(2017, 4, 23, 13, 35)
-        }, {
-            id: 6,
-            text: "Install New Database",
-            startDate: new Date(2017, 4, 24, 9, 45),
-            endDate: new Date(2017, 4, 24, 11, 15)
-        }, {
-            id: 7,
-            text: "Upgrade Personal Computers",
-            startDate: new Date(2017, 4, 24, 15, 15),
-            endDate: new Date(2017, 4, 24, 16, 30)
-        }, {
-            id: 8,
-            text: "Customer Workshop",
-            startDate: new Date(2017, 4, 25, 11, 0),
-            endDate: new Date(2017, 4, 25, 12, 0),
-        }, {
-            id: 9,
-            text: "Prepare 2015 Marketing Plan",
-            startDate: new Date(2017, 4, 25, 11, 0),
-            endDate: new Date(2017, 4, 25, 13, 30)
-        }, {
-            id: 10,
-            text: "Create Icons for Website",
-            startDate: new Date(2017, 4, 26, 10, 0),
-            endDate: new Date(2017, 4, 26, 11, 30)
-        }, {
-            id: 11,
-            text: "Upgrade Server Hardware",
-            startDate: new Date(2017, 4, 26, 14, 30),
-            endDate: new Date(2017, 4, 26, 16, 0)
-        }, {
-            id: 12,
-            text: "Submit New Website Design",
-            startDate: new Date(2017, 4, 27, 16, 30),
-            endDate: new Date(2017, 4, 27, 18, 0)
-        }, {
-            id: 13,
-            text: "Launch New Website",
-            startDate: new Date(2017, 4, 27, 12, 20),
-            endDate: new Date(2017, 4, 27, 14, 0)
-        }
-    ];
-
-    const createDataSource = (list = defaultData) => {
-        return new DataSource({
-            store: {
-                type: "array",
-                key: "id",
-                data: [...list]
-            }
-        });
-    };
-
-    QUnit.test("Scheduler should render appointments only for appointments that need redraw", function(assert) {
-        const dataSource = createDataSource();
-        const scheduler = createScheduler("week", dataSource);
-
-        markAppointments(scheduler);
-        dataSource.store().push([
-            { type: "update", key: 0, data: { text: "updated-1" } },
-            { type: "update", key: 1, data: { text: "updated-2" } }
-        ]);
-        dataSource.load();
-        assert.equal(2, getUnmarkedAppointments(scheduler).length, "Should rendered only two updated appointments");
-
-        markAppointments(scheduler);
-        dataSource.store().push([{ type: "insert", data: {
-            id: 15,
-            text: "Fake",
-            startDate: new Date(2017, 4, 27, 15, 30),
-            endDate: new Date(2017, 4, 27, 16, 30)
-        } }]);
-        dataSource.load();
-        assert.equal(1, getUnmarkedAppointments(scheduler).length, "Should rendered only inserted appointment");
-
-        markAppointments(scheduler);
-        dataSource.store().remove(0);
-        dataSource.load();
-        assert.equal(0, getUnmarkedAppointments(scheduler).length, "Html element should removed and should not redrawing another appointments");
-    });
-
-    QUnit.test("Scheduler should render only necessary appointments in crossing appointments case", function(assert) {
-        const dataSource = createDataSource();
-        const scheduler = createScheduler("week", dataSource);
-
-        markAppointments(scheduler);
-        dataSource.store().push([{ type: "insert", data: {
-            id: 14,
-            text: "Fake_key_14",
-            startDate: defaultData[0].startDate,
-            endDate: defaultData[0].endDate
-        } }]);
-        dataSource.load();
-        assert.equal(2, getUnmarkedAppointments(scheduler).length, "Should rendered inserted appointment and update appointment");
-
-        markAppointments(scheduler);
-        dataSource.store().push([{ type: "insert", data: {
-            id: 15,
-            text: "Fake_key_15",
-            startDate: defaultData[1].startDate,
-            endDate: defaultData[1].endDate
-        } }]);
-        dataSource.load();
-        assert.equal(2, getUnmarkedAppointments(scheduler).length, "Should rendered inserted appointment and 2 updated appointment");
-
-        markAppointments(scheduler);
-        dataSource.store().remove(15);
-        dataSource.load();
-        assert.equal(1, getUnmarkedAppointments(scheduler).length, "Should rendered only two updated appointments");
-    });
-
-    QUnit.test("Scheduler should throw onAppointmentRendered event only for appointments that need redraw", function(assert) {
-        const dataSource = createDataSource();
-        const fakeHandler = {
-            onAppointmentRendered: () => { }
-        };
-        const renderedStub = sinon.stub(fakeHandler, "onAppointmentRendered");
-        createScheduler("week", dataSource, { onAppointmentRendered: fakeHandler.onAppointmentRendered });
-
-        renderedStub.reset();
-        dataSource.store().push([{ type: "insert", data: {
-            id: 14,
-            text: "Fake_key_14",
-            startDate: new Date(2017, 4, 21, 15, 0),
-            endDate: new Date(2017, 4, 21, 15, 30),
-        } }]);
-        dataSource.load();
-        assert.equal(renderedStub.callCount, 1, "Should throw one call onAppointmentRendered event");
-
-        renderedStub.reset();
-        dataSource.store().remove(14);
-        dataSource.load();
-        assert.equal(renderedStub.callCount, 0, "Should not throw onAppointmentRendered event");
-
-        renderedStub.reset();
-        dataSource.store().push([{ type: "insert", data: {
-            id: 15,
-            text: "Fake_key_15",
-            startDate: defaultData[0].startDate,
-            endDate: defaultData[0].endDate
-        } }]);
-        dataSource.load();
-        assert.equal(renderedStub.callCount, 2, "Should throw two call onAppointmentRendered event");
-
-        renderedStub.reset();
-        dataSource.store().push([
-            { type: "update", key: 0, data: { text: "updated-1" } },
-        ]);
-        dataSource.load();
-
-        assert.equal(renderedStub.callCount, 1, "Should throw one call onAppointmentRendered event");
-    });
-
-    QUnit.test("Scheduler should render appointments only for appointments that need redraw in Month view", function(assert) {
-        const dataSource = createDataSource();
-        const scheduler = createScheduler("month", dataSource);
-
-        markAppointments(scheduler);
-        dataSource.store().push([
-            { type: "update", key: 0, data: { text: "updated-1" } },
-            { type: "update", key: 1, data: { text: "updated-2" } }
-        ]);
-        dataSource.load();
-        assert.equal(2, getUnmarkedAppointments(scheduler).length, "Should rendered only two updated appointments");
-
-        markAppointments(scheduler);
-        dataSource.store().push([{ type: "insert", data: {
-            id: 15,
-            text: "Fake",
-            startDate: new Date(2017, 4, 28, 15, 30),
-            endDate: new Date(2017, 4, 28, 16, 30)
-        } }]);
-        dataSource.load();
-        assert.equal(1, getUnmarkedAppointments(scheduler).length, "Should rendered only inserted appointment");
-
-        markAppointments(scheduler);
-        dataSource.store().remove(0);
-        dataSource.load();
-
-        // TODO: in future this case should be optimized - redraw in this case can escape
-        assert.equal(1, getUnmarkedAppointments(scheduler).length, "Should rendered only one appointment");
-    });
-
-    QUnit.test("Scheduler should render appointments only for appointments that need redraw. Use scheduler API", function(assert) {
-        const scheduler = createScheduler("week", defaultData);
-
-        markAppointments(scheduler);
-        scheduler.instance.updateAppointment(defaultData[0], { text: "updated" });
-        assert.equal(1, getUnmarkedAppointments(scheduler).length, "Should rendered only one appointment");
-
-        markAppointments(scheduler);
-        scheduler.instance.updateAppointment(defaultData[9], { text: "updated" });
-        assert.equal(1, getUnmarkedAppointments(scheduler).length, "Should rendered only one appointment from intersecting appointments");
-
-        markAppointments(scheduler);
-        scheduler.instance.deleteAppointment(defaultData[0]);
-        assert.equal(0, getUnmarkedAppointments(scheduler).length, "Nothing should be redrawing");
-    });
-
-    QUnit.test("Scheduler should re-render all appointments in Agenda view case", function(assert) {
-        const dataSource = createDataSource();
-        const scheduler = createScheduler("agenda", dataSource);
-
-        markAppointments(scheduler);
-        dataSource.store().push([
-            { type: "update", key: 8, data: { text: "updated-1" } },
-            { type: "update", key: 10, data: { text: "updated-2" } }
-        ]);
-        dataSource.load();
-        assert.equal(scheduler.appointments.getAppointmentCount(), getUnmarkedAppointments(scheduler).length, "Should rendered all appointments");
-
-        markAppointments(scheduler);
-        dataSource.store().push([{ type: "insert", data: {
-            id: 15,
-            text: "Fake",
-            startDate: new Date(2017, 4, 27, 15, 30),
-            endDate: new Date(2017, 4, 27, 16, 30)
-        } }]);
-        dataSource.load();
-        assert.equal(scheduler.appointments.getAppointmentCount(), getUnmarkedAppointments(scheduler).length, "Should rendered all appointments");
-    });
-
-    QUnit.test("Scheduler should re-render appointments in Agenda view, if data source loading data", function(assert) {
-        const items = [
-            { id: 0, startDate: new Date(2017, 4, 25, 9), endDate: new Date(2017, 4, 25, 9, 30), text: "a" },
-            { id: 1, startDate: new Date(2017, 4, 27, 15), endDate: new Date(2017, 4, 27, 15, 30), text: "b" }
-        ];
-
-        const dataSource = {
-            store: new CustomStore({
-                key: "id",
-                load: () => items,
-                update: (key, values) => items[parseInt(key)] = values
-            })
-        };
-        const scheduler = createScheduler("agenda", dataSource);
-        assert.equal(scheduler.appointments.getAppointmentCount(), 2, "Should render 2 appointments");
-        markAppointments(scheduler);
-
-        scheduler.appointments.click();
-        scheduler.tooltip.clickOnItem();
-        scheduler.appointmentForm.setSubject("new text");
-        scheduler.appointmentPopup.clickDoneButton();
-
-        assert.equal(scheduler.appointments.getAppointmentCount(), 2, "Should render 2 appointments");
-        assert.equal(scheduler.appointments.getAppointmentCount(), getUnmarkedAppointments(scheduler).length, "Should re-rendered all appointments");
-    });
-});
 
 QUnit.module("LayoutManager", moduleOptions);
